Change header in default Fingerprint generator
In default FingerprintGenerator, replace ACCEPT header with ACCEPT-LANGUAGE header (when generating Fingerprint):
Using the ACCEPT header field for the default Fingerprint generator will cause problems for applications that use different media types depending on requests. For example, you may have a controller that accepts "application/json, text/javascript, /; q=0.01" for RESTful requests, and "text/html,application/xhtml+xml,application/xml;q=0.9" for HTML requests. Although the user can implement his/her own generator in this case, it will be good to avoid the "ACCEPT" header for the default generator (so as to remove a potential complication for new users who, for example, use such a mixture of REST and HTML in their API). ACCEPT-LANGUAGE, on the other hand, will be least likely to be different across request types.