Add email support to Twitter authentication
Earlier in 2015 twitter added the ability for applications to retrieve
user emails. Applications have to be specifically whitelisted in order
to do this, and an additional token has to be sent along with the request
to verify_credentials in order for the email to be returned.

More information:
  https://dev.twitter.com/rest/reference/get/account/verify_credentials

In addition, exclude twitter status from API call

The verify_credentials API by default returns the current status of
the user being requested. Since this feature is not used as part of
Silhouette, disabling in the request parameters will save precious
pico-seconds.