Use None as default value for Option based settings values
If we use an other default value as `None` for `Option` based setting keys, then we do not have the possibility to set `None` as value in the Play configuration. Because if we omit the config property or if we use `null` as value, then the Typesafe config implementation will use the default value for this property.