Move persistence based API into a separate project
It's often misleading that the persistence based API is integrated into the core project, because it's not really required by the core. The provides API is only one possible implementation which helps to start a new application quicker. Therefore this pull request moves the persistence based API into a separate project and provides a in-memory implementation for the auth info implementations. It's now also possible to provide different drop-in replacements for the persistence layer instead of creating a separate seed template for every persistence type(reactivemongo, slick, ...).