Fix broken tests introduced in ad7bcdc5
Since the introduction of API URL override in ad7bcdc5 the build was
inevitably broken because of a wrong Mockito validate statement.
Travis did not detect the breakage because of the missing "-e" option
on the bash execution, thus the final "echo" was shallowing the exit
code of the test.

This fixes the build ci script *AND* the Mockito validation statement.

Additionally all the waits in tests were waiting at most 1 second which
could be not enough for the test to complete, especially on Travis 
infrastructure.