Remove the persistence in-memory modules (#479)
It's not really necessary to provide in-memory modules for the persistence layer, because the persistence layer module provides already a `InMemoryAuthInfoDAO` which can be used to bind in-memery DAOs for every auth info. The binding for an auth info DAO is only a one-liner, and therefore the maintenance of the module code does not stand in proportion to the necessary effort.