Allow to pass user state with the social state handlers (#503)
Adds the ability to let the user pass context state to the OAuth2 providers. The main application for this is redirection to the same page that a user originally requested but was not authenticated for; in OAuth2 there is a separate redirect_url (aka callbackURL) param, but most providers restrict that to a preset whitelist of URLs in order to combat CSRF attacks; the best practice recommended by e.g. Google is to have state be a structured field that includes both the CSRF token and "any other information needed to recover the context when the user returns to your application, e.g. the starting URL."