State handler uses wrong result if a user state item is used (#531)
The problem is that the `getOrElse` method returns the result passed to the `publish` method and not the accumulated result. So if a non publishable state items comes after a publishable state item, then it overrides the result with original result and therefore throws away all changes made by the publishable item.